Tạm thời sửa những phần này

+++ app/src/main/java/com/haunp/mybookstore/data/database/dao/OrderDao.kt
@@ -11,7 +11,7 @@ import com.haunp.mybookstore.domain.entity.OrderEntity
 @Dao
 interface OrderDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertOrder(order: OrderEntity): Int
+    suspend fun insertOrder(order: OrderEntity): Long
 
     @Query("SELECT * FROM orders WHERE userId = :userId")
     suspend fun getOrdersByUserId(userId: Int): List<OrderEntity>
